Make `arbitrary` an optional dependency
Shuffle around a few things in a few locations for this.
