+++ cranelift/assembler-x64/Cargo.toml
@@ -7,7 +7,7 @@ edition.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-arbitrary = { version = "1.3.2", features = ["derive"] }
+arbitrary = { workspace = true, features = ["derive"], optional = true }
 
 [dev-dependencies]
 arbtest = "0.3.1"
@@ -21,3 +21,6 @@ pedantic = "warn"
 module_name_repetitions = { level = "allow", priority = 1 }
 similar_names = { level = "allow", priority = 1 }
 wildcard_imports = { level = "allow", priority = 1 }
+
+[features]
+arbitrary = ['dep:arbitrary']
