+++ cranelift/assembler-x64/fuzz/Cargo.toml
@@ -10,7 +10,7 @@ cargo-fuzz = true
 
 [dependencies]
 libfuzzer-sys = { workspace = true }
-cranelift-assembler-x64 = { path = ".." }
+cranelift-assembler-x64 = { path = "..", features = ['arbitrary'] }
 capstone = { workspace = true }
 arbitrary = { workspace = true, features = ['derive'] }
